syslog input use syslog parser test case
